app/requirements.txt
requests==2.25.1
flask==1.1.2
python-dotenv==0.15.0
PyYAML==5.4.1